requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.5
openai==1.3.5
faiss-cpu==1.7.4
pypdf2==3.0.1
python-multipart==0.0.6
sentence-transformers==2.2.2
chromadb==0.4.18
python-dotenv==1.0.0